--- a/c3bottles/__init__.py
+++ b/c3bottles/__init__.py
@@ -8,7 +8,6 @@ from flask_babel import Babel
 from flask_bcrypt import Bcrypt
 from flask_compress import Compress
 from flask_login import LoginManager
-from flask_migrate import Migrate
 from flask_sqlalchemy import SQLAlchemy
 from flask_wtf import CSRFProtect
 
@@ -34,7 +33,6 @@ if "pytest" not in sys.modules:
 Compress(app)
 
 db = SQLAlchemy(app, session_options={"autoflush": False})
-migrate = Migrate(app, db)
 lm = LoginManager(app)
 bcrypt = Bcrypt(app)
 csrf = CSRFProtect(app)