--- a/c3bottles/model/drop_point.py
+++ b/c3bottles/model/drop_point.py
@@ -100,6 +100,7 @@ class DropPoint(db.Model):
             raise ValueError(*errors)
 
         db.session.add(self)
+        db.session.commit()
 
     def remove(self, time=None):
         """