--- a/c3bottles/model/location.py
+++ b/c3bottles/model/location.py
@@ -99,6 +99,7 @@ class Location(db.Model):
             raise ValueError(*errors)
 
         db.session.add(self)
+        db.session.commit()
 
     @property
     def description_with_level(self):