--- a/c3bottles/model/report.py
+++ b/c3bottles/model/report.py
@@ -3,6 +3,7 @@ from datetime import datetime
 from flask_babel import lazy_gettext
 
 from c3bottles import db
+from c3bottles.lib import metrics
 from c3bottles.model import drop_point
 
 
@@ -79,6 +80,10 @@ class Report(db.Model):
         db.session.add(dp)
         db.session.add(self)
         db.session.commit()
+
+        metrics.report_count.labels(
+            state=self.state, category=self.dp.category.metrics_name
+        ).inc()
 
     def get_weight(self):
         """Get the weight (i.e. significance) of a report.