--- a/requirements/production.txt
+++ b/requirements/production.txt
@@ -6,7 +6,6 @@ Flask-Babel>=0.12.2
 Flask-Bcrypt>=0.7.1
 Flask-Compress>=1.4.0
 Flask-Login>=0.4.1
-Flask-Migrate>=2.5.2
 Flask-SQLAlchemy>=2.4.1
 Flask-WTF>=0.14.2
 # remember to update pillow in docker.txt as well!