First I have removed Flask-Migrate/Alembic as it is not really needed and does not work if c3bottles has been installed like we describe in the documentation. In addition, the second migration fails for SQLite which I use a lot for testing, so I simply dropped Flask-Migrate. Our models and the setup are so simple that we do not need it anyway.

Apart from small cosmetic cleanups, the main change is adding a state column to the drop point table. Before, when querying all drop points and their states (like for the map or the list), that would trigger one select for all the drop points and then another two queries for every single drop point to calculate its state. When adding the state to the database explictly, only one single query is needed.

Before:

```
In [5]: %timeit [dp.last_state for dp in drop_point.DropPoint.query.all()]
11.4 ms ± 41 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

After:

```
In [10]: %timeit [dp.last_state for dp in drop_point.DropPoint.query.all()]
466 µs ± 2.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

This is for a small dataset and gets even better, the more drop points and reports there are.

By doing something similar with the calculation of visit priority, we probably will be able to increase performance even further.

In addition, now that we set the state explicitly for every drop point, we can add this to the Prometheus metrics properly via labels. *looks at @J-Ha*