@MichiK 2019-10-13
There is more:

* [ ] location (this needs changing the location constructor so that we can construct locations in memory, serialize them and add them to the drop point table)
* [ ] count and weight of new reports (since the last visit) for priority calculation
* [ ] total report and visit count

The goal would be to be able to query all relevant properties of drop points (apart from details of past reports, visits and locations which are only shown in the history for a single drop point) from one single request to the database for all drop points at once.